+++ gazebo/gui/model/JointMaker_TEST.cc
@@ -454,5 +454,34 @@ void JointMaker_TEST::Selection()
   delete mainWindow;
 }
 
+/////////////////////////////////////////////////
+void JointMaker_TEST::JointMaterial()
+{
+  this->Load("worlds/empty.world");
+
+  gui::JointMaker *jointMaker = new gui::JointMaker();
+
+  // all currently supported joint types.
+  std::vector<std::string> jointTypes;
+  jointTypes.push_back("revolute");
+  jointTypes.push_back("revolute2");
+  jointTypes.push_back("prismatic");
+  jointTypes.push_back("ball");
+  jointTypes.push_back("universal");
+  jointTypes.push_back("screw");
+  jointTypes.push_back("gearbox");
+
+  // verify joint materials are not empty and they are all unique
+  std::set<std::string> jointMaterials;
+  for (auto &j : jointTypes)
+  {
+    std::string mat = jointMaker->GetJointMaterial(j);
+    QVERIFY(mat != "");
+    QVERIFY(jointMaterials.find(mat) == jointMaterials.end());
+    jointMaterials.insert(mat);
+  }
+  delete jointMaker;
+}
+
 // Generate a main function for the test
 QTEST_MAIN(JointMaker_TEST)