+++ gazebo/gui/model/SchematicViewWidget.hh
@@ -59,21 +59,33 @@ namespace gazebo
       public: void RemoveNode(const std::string &_node);
 
       /// \brief Check if a node exists in the scene in the widget.
-      /// \param[in] _name Name of the name.
+      /// \param[in] _name Name of the node.
       /// \return True if the node exists.
       public: bool HasNode(const std::string &_name) const;
 
       /// \brief Add an edge to the scene in the widget
       /// \param[in] _id Unique id of edge.
       /// \param[in] _name Name of edge.
+      /// \param[in] _name Type of edge.
       /// \param[in] _parent Name of parent node.
       /// \param[in] _child Name of child node.
       public: void AddEdge(const std::string &_id, const std::string &_name,
-          const std::string &_parent, const std::string &_child);
+          const std::string &_type, const std::string &_parent,
+          const std::string &_child);
 
       /// \brief Remove an edge from the scene in the widget
       /// \param[in] _id Unique id of edge.
       public: void RemoveEdge(const std::string &_id);
+
+      /// \brief Update an edge in the scene
+      /// \param[in] _id Unique id of edge.
+      /// \param[in] _name Name of edge.
+      /// \param[in] _type Type of edge.
+      /// \param[in] _parent Name of parent node.
+      /// \param[in] _child Name of child node.
+      public: void UpdateEdge(const std::string &_id, const std::string &_name,
+          const std::string &_type, const std::string &_parent,
+          const std::string &_child);
 
       /// \brief Get number of nodes in the scene.
       /// \return Number of nodes.