+++ Changelog.md
@@ -1,3 +1,9 @@
+## Gazebo 7.0
+
+1. Model editor updates
+    1. Color-coded edges in Schematic View to match joint color.
+        * [Pull request #1781](https://bitbucket.org/osrf/gazebo/pull-request/1781)
+
 ## Gazebo 6.0
 
 1. Added magnetometer sensor. A contribution from Andrew Symington.